refactor frontend app.js into components
