+++ frontend/src/components/TabsContainer.js
@@ -1,0 +1,41 @@
+import React, { useState } from 'react';
+import { Tabs, Tab, Box } from '@mui/material';
+
+function TabPanel(props) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box p={3}>{children}</Box>}
+    </div>
+  );
+}
+
+export default function TabsContainer({ tabs }) {
+  const [tabValue, setTabValue] = useState(0);
+
+  const handleTabChange = (event, newValue) => {
+    setTabValue(newValue);
+  };
+
+  return (
+    <>
+      <Tabs value={tabValue} onChange={handleTabChange} aria-label="tabs">
+        {tabs.map((tab, index) => (
+          <Tab key={index} label={tab.label} />
+        ))}
+      </Tabs>
+      {tabs.map((tab, index) => (
+        <TabPanel key={index} value={tabValue} index={index}>
+          {tab.content}
+        </TabPanel>
+      ))}
+    </>
+  );
+}
